Implements Basic Setup for Web Scraping
Adds the /Webscraper/GatherJobData endpoint which calls the GetWebsiteData() method of every IWebScraper implementation. Thanks to the ScraperFactory, additional Website Scrapers can easily be added by creating a class that inherits from IWebScraper and adding that class to Dependency Injection in the ScraperFactoryConfiguration class.

Also adds a wrapper for the Html Agility Pack library (used for web scraping) to improve testability.